Update eat methode and apply to all piece
