+++ chess_prog/Pieces/bishop.py
@@ -21,16 +21,3 @@ class Bishop(Piece):
         self.pos.x = x
         self.pos.y = y
         return True
-
-    def eat(self, x: int, y: int, all_piece: list):
-        temp = None
-        for i in all_piece:
-            if i.pos.same_pos(Position(x, y)) and self.white != i.white:
-                temp = i
-        if not temp:
-            return False
-        return_value = self.move(x, y, all_piece)
-        if return_value:
-            all_piece.remove(temp)
-            return True
-        return False
